Changed mqtt_publish_interval to int #4
*Fix type inconsistency by changing mqtt_publish_interval to int

+++ Software/Firmware/src/connectivity.cpp
@@ -11,7 +11,7 @@
 #include "time.h"
 #include "hardware.h"
 
-unsigned long mqtt_publish_interval = 500; //MQTT Publish Interval
+int mqtt_publish_interval = 500; //MQTT Publish Interval
 
 AsyncWebServer server(80);
